pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptshield"
version = "0.1.0"
description = "PromptShield – Automated LLM Security Gate"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "PromptShield"}]
dependencies = [
  "PyYAML>=6.0",
  "httpx>=0.25.0",
  "google-adk>=0.3.0",
  "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
full = [
  "garak>=0.13.0",
  "pyrit>=0.10.0",
  "llm-guard>=0.3.0",
]

[project.scripts]
promptshield = "promptshield.cli:main"
promptguard = "promptguard.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
